requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
minio
pydantic-settings
httpx
openai
numpy
scikit-learn
python-dateutil
python-multipart
cryptography
sentence-transformers
torch
python-jose[cryptography]
einops
# Document loaders (Phase 1.2)
PyPDF2>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
